add more relevant picture.png
Just in case the user decides to open the picture, we're now
including a more revelant picture (qubes architecture).

The picture was obtained by xfce4-screenshooter of the qubes
website. But it would be advisable to replace the image with one
generated by a qubes team member at the time of review.